[Feat] Revision System + Roaster Revision Implementation (#113)
- Create Revision for Roaster API, UI 
- Update Revision for Roaster API, UI
- Approve Revision for Roaster API (UI implemented later)
- Reject Revision for Roaster API (UI implemented later)
- View Revisions of Authenticated User API, UI
- View Revision Diff for Roaster API, UI
- View Committed Revision History for Roaster API, UI
- View Revision Snapshot API
- View Pending Revisions as Moderator API (UI implemented later)
- View Committed Revision Contributions for User API, UI

+++ Backend/Common/Params/RoasterParams.cs
@@ -11,14 +11,14 @@ public class RoasterParams : PaginationParams, IValidatableObject
   public string? Address { get; set; }
 
   // PostGIS search
-  [Range(typeof(float), "-90", "90", ErrorMessage = "Latitude must be between -90 and 90.")]
-  public float? Lat { get; set; }
+  [Range(typeof(double), "-90", "90", ErrorMessage = "Latitude must be between -90 and 90.")]
+  public double? Lat { get; set; }
 
-  [Range(typeof(float), "-180", "180", ErrorMessage = "Latitude must be between -90 and 90.")]
-  public float? Long { get; set; }
+  [Range(typeof(double), "-180", "180", ErrorMessage = "Latitude must be between -90 and 90.")]
+  public double? Long { get; set; }
 
-  [Range(typeof(float), "0.1", "15000", ErrorMessage = "Radius must be between 0.1km and 15000km.")]
-  public float? Radius { get; set; }
+  [Range(typeof(double), "0.1", "15000", ErrorMessage = "Radius must be between 0.1km and 15000km.")]
+  public double? Radius { get; set; }
 
   public IEnumerable<ValidationResult> Validate(ValidationContext _)
   {
